Added Not Found Exception

+++ src/main/java/com/example/jpatest/service/BrandServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.jpatest.exception.NotFoundException;
 import com.example.jpatest.model.Brand;
 import com.example.jpatest.repository.BrandRepository;
 
@@ -25,6 +26,7 @@ public class BrandServiceImpl implements BrandService {
 	
 	@Override
 	public void delete(Brand brand) {
+		if(!brandRepository.existsById(brand.getId())) throw new NotFoundException("Brand not found");
 		brandRepository.deleteById(brand.getId());
 	}
